Add files via upload
I wrote some python code that used if, elif, and else statements to create a decision-making structure in my program. The if statement allows me to specify a condition to be checked, and if the condition is true, a block of code is executed. The elif statement allows me to specify additional conditions to be checked if the previous conditions are not true. The else statement is used to specify a block of code to be executed if none of the conditions are true. These statements allowed me to create a branching structure in my code, allowing me to choose which code to execute based on the evaluated conditions.