Add files via upload
I wrote some python code that defined functions, which are blocks of reusable code that can be called multiple times from different parts of my program. To define a function, I used the def keyword followed by the function name and a set of parentheses that may include parameters. I then specified the block of code to be executed when the function is called, using indentation to indicate the start and end of the code block.

I can define default values for the parameters in my function definition, which allows me to use the function with fewer arguments than the number of parameters. I can also specify keyword arguments in the function call, which allows me to specify the values for specific parameters by name.

Overall, defining functions allowed me to modularize my code and reuse the same block of code multiple times in my program.