Add files via upload
I wrote some python code that defined functions and incorporated loops to perform repetitive tasks. Functions are blocks of reusable code that can be called multiple times from different parts of my program. To define a function, I used the def keyword followed by the function name and a set of parentheses that may include parameters. I then specified the block of code to be executed when the function is called, using indentation to indicate the start and end of the code block.

I also used loops, such as for loops and while loops, to repeat a block of code multiple times. The for loop allows me to iterate over a sequence of elements, such as a list or a string, and execute a block of code for each element. The while loop allows me to specify a condition that is checked before each iteration, and the loop continues to execute as long as the condition is true.

I incorporated loops into my functions by writing the loop within the function definition, allowing me to perform repetitive tasks each time the function is called. This allowed me to modularize my code and reuse the same block of code multiple times in my program, while also performing repetitive tasks as needed.