Add files via upload
I wrote some python code that defined functions and incorporated lists to store and manipulate data. Functions are blocks of reusable code that can be called multiple times from different parts of my program. To define a function, I used the def keyword followed by the function name and a set of parentheses that may include parameters. I then specified the block of code to be executed when the function is called, using indentation to indicate the start and end of the code block.

I also used lists to store and manipulate data in my functions. Lists are ordered collections of items that can be of any data type, and they allow me to store multiple values in a single variable. I can access and modify the elements of a list using indexing and slicing, and I can use various methods and functions to manipulate the list, such as appending, inserting, removing, and sorting elements.

I incorporated lists into my functions by passing lists as arguments to the function, or by defining lists within the function and manipulating them as needed. This allowed me to store and manipulate data within my functions and use that data to perform specific tasks.