Add files via upload
I wrote some python code that used lists to store and manipulate data. Lists are ordered collections of items that can be of any data type, and they allow me to store multiple values in a single variable. I used indexing and slicing to access and modify the elements of a list, and I used various methods and functions to manipulate the list, such as appending, inserting, removing, and sorting elements.

I also used loops, such as for loops and while loops, to iterate over the elements of a list and perform specific tasks on each element. This allowed me to efficiently process large amounts of data stored in lists.

Overall, list manipulation allowed me to store and manipulate multiple values in a single variable, and to perform tasks on those values in an efficient and organized manner.