Add files via upload
I wrote some python code that defined classes, which are templates for creating objects. A class defines the attributes and behaviors of an object, and an object is an instance of a class.

To define a class, I used the class keyword followed by the class name and a colon, and then specified the class attributes and behaviors within the class definition using indentation. I defined class attributes by assigning values to variables within the class definition, and I defined class behaviors using methods, which are functions defined within the class.

I created an object by calling the class name and passing any necessary arguments to the class constructor, which is a special method used to initialize the object. I can access and modify the object's attributes and behaviors using dot notation, which involves specifying the object name followed by a dot and the attribute or method name.

Overall, using classes allowed me to create objects that have specific attributes and behaviors, and to reuse those objects in my program as needed.